[fix] change github auth to personal access token, read more
The problem with this deployment was regarding the access permission. Since the GSP repo is hosted on an Organization account, it requires a different setup.

I changed the code as you can see in the pages.yml file from github access token to PAT.
Then I generated a new Personal Accees Token (PAT) by going to my profile > settings > developer settings > access tokens (classic) and selected Repo and Actions options for permissions.

Then I headed to GSP repo settings > secrets > created a secret named PAT, filled in the value and that's it.